requirements.txt
groq
python-dotenv
pdfplumber
python-docx
fastapi
uvicorn
pydantic
python-multipart
websockets
pytesseract
pdf2image
PyMuPDF
Pillow
spacy
